requirements.txt
torch
torchvision
ultralytics
gradio
pillow
python-multipart
uvicorn
fastapi
